Fix: libcrmcluster: use uint64_t type for corosync ringid (membership  id) when updating node state

+++ include/crm/cluster/internal.h
@@ -317,7 +317,7 @@ gboolean init_cs_connection_once(crm_cluster_t * cluster);
 crm_node_t *crm_update_peer_proc(const char *source, crm_node_t * peer,
                                  uint32_t flag, const char *status);
 crm_node_t *crm_update_peer_state(const char *source, crm_node_t * node,
-                                  const char *state, int membership);
+                                  const char *state, uint64_t membership);
 
 void crm_update_peer_uname(crm_node_t *node, const char *uname);
 void crm_update_peer_expected(const char *source, crm_node_t * node, const char *expected);
